+++ estimators/bandits/gaussian.py
@@ -14,7 +14,7 @@ class Interval(base.Interval):
         self.weighted_reward = 0
         self.weighted_reward_sq = 0
 
-    def add_example(self, p_log: float, r: float, p_pred: float, count: float = 1.0) -> None:
+    def add_example(self, p_log: float, r: float, p_pred: float, count: int = 1) -> None:
         assert count == 1.0, "need to explicitly model the pdrop generatively in order to prevent misleading confidence interval widths"
         self.examples_count += count
         w = p_pred/p_log
